pyproject.toml
[project]
name = "kg-workflow"
version = "0.1.0"
description = "A project with utilitiy functions for working with OWL based knowledge graphs."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "exceptiongroup==1.3.0",
    "iniconfig==2.1.0",
    "isodate==0.7.2",
    "numpy==2.2.6",
    "packaging==25.0",
    "pandas==2.3.3",
    "pandas-stubs==2.3.2.250926",
    "pluggy==1.6.0",
    "pygments==2.19.2",
    "pyparsing==3.2.5",
    "pytest==8.4.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "rdflib==7.2.1",
    "six==1.17.0",
    "tomli==2.2.1",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.scripts]
sparql-select = "kgworkflow.sparql_select:main"
solve-zebra = "kgworkflow.main:main"


